Dynamically load parsing so users can roll their own
