+++ app.js
@@ -14,116 +14,6 @@ const app = new App({
   appToken: process.env.SLACK_APP_TOKEN
 });
 
-/*
- * returns an array of block elements representing a decision log entry
- * that can be individually written to response using message.blocks.push().
- *
- * takes an `adrFile` object that resembles:
- * {
- *  name: <filename>,
- *  object: {
- *    text: <file contents>
- *  }
- * }
- */
-function toBlockFormat(adrFile) {
-
-  let block = [
-    {
-      type: "divider"
-    },
-  ];
-
-  const adrJsonObj = adrFile.data;
-
-  // push the adr title
-  if(adrJsonObj.title) {
-    block.push({
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: `*Problem:* <${adrFile.githubUrl}|${adrJsonObj.title}>`,
-      },
-			accessory: {
-				type: "button",
-				text: {
-					type: "plain_text",
-					text: "List PRs",
-				},
-				value: adrFile.name,
-				action_id: "list prs action"
-			}
-    });
-  }
-
-  if(adrJsonObj["Problem Description"]) {
-    block.push({
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: adrJsonObj["Problem Description"],
-      },
-    });
-  }
-
-  if(adrJsonObj["Accepted Solution"]) {
-    block.push({
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: "*Accepted Solution*",
-      },
-    });
-
-
-    block.push({
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: adrJsonObj["Accepted Solution"],
-      },
-    });
-
-  }
-
-  // if there is frontmatter, add that as a Context block in slack message
-  if (adrJsonObj.frontmatter && (
-    adrJsonObj.frontmatter.status
-    || adrJsonObj.frontmatter.committed
-    || adrJsonObj.frontmatter["decide-by"]
-    || adrJsonObj.frontmatter["review-by"])
-    ) {
-    
-    // create reader friendly versions of the frontmatter properties
-    const labels = {
-      status: "Status",
-      "committed-on": "Committed On",
-      "decide-by": "Decide By",
-      "review-by": "Review By",
-      impact: "Impact",
-    };
-    
-    const elements = [];
-    for (const property in labels) {
-      
-      // add a context element if there is a matching property in the frontmatter
-      if (adrJsonObj.frontmatter[property]) {
-        elements.push(
-          {
-            type: "mrkdwn",
-            text: `\`${labels[property]}: ${adrJsonObj.frontmatter[property]}\``,            
-          }
-        );
-      }
-    }
-    block.push({
-      type: "context",
-      elements: elements
-    });
-  }
-
-  return block;
-}
 
 /*
  * respond to a user clicking the "View PRs" button by opening a modal with
@@ -298,8 +188,10 @@ app.command("/decision", async ({ command, ack, respond }) => {
 
         for (const adrFile of adrFiles) {
 
+          const blockFormatter = await import(process.env.ADR_TO_BLOCK_FORMATTER || "./lib/blockFormatter.js");
+
           // convert adr file to Slack block format and push it to the message body
-          const blocks = toBlockFormat(adrFile);
+          const blocks = blockFormatter.toBlockFormat(adrFile);
 
           blocks.forEach(block => {
             message.blocks.push(block);
