+++ lib/adrs.js
@@ -2,7 +2,6 @@ import { Octokit } from "octokit";
 import {unified} from 'unified'
 import remarkParse from 'remark-parse'
 import remarkFrontmatter from 'remark-frontmatter'
-import YAML from 'yaml'
 
 // establish github connection
 const octokit = new Octokit({ auth: process.env.GITHUB_TOKEN });
@@ -87,8 +86,8 @@ export async function getAdrFiles(options) {
         .use(remarkFrontmatter)
         .parse(text);
     
-    
-      adrEntry.data = adrToJSON(adrAST);
+      const parser = await import(process.env.ADR_PARSER || "./adrParser.js");
+      adrEntry.data = parser.adrToJSON(adrAST);
     
       if (checkFilter(adrEntry.data.frontmatter,options)){
         adrFiles.push(adrEntry);
@@ -167,42 +166,7 @@ function checkFilter(frontmatter, options) {
 
 
 
-/*
- * takes an AST and a section header, and finds the associated text under that header.
- */
 
-function adrToJSON(ast) {
-  
-  // 'children' is an array of markdown blocks, such as headers, paragraphs, lists etc
-  const {children} = ast;
-
-  let jsonObj = {};
-
-  // look for YAML frontmatter
-  if (children[0].type == "yaml") {
-    jsonObj.frontmatter = YAML.parse(children[0].value);
-  }
-
-  // the AST should start with the YAML frontmatter, and should have a title section next
-  if (children.length > 1) {
-  // get the title - this is at position [1] and has depth of 1
-    jsonObj.title = children[1].children[0].value;
-
-    // The next paragraph is the text we are after
-
-    let i = 0;
-
-    while (i < children.length) {
-        if (children[i].type == "heading" && children[i].depth == 2) {
-        // found a section header
-        jsonObj[children[i].children[0].value] = children[i+1].children[0].value;
-        }
-
-        i++;
-    }
-  }
-  return jsonObj;
-}
   
 
 // returns the list of pull requests.
